Solución en un solo fichero para el problema 32 y 33. Perdemos los tests unitarios de esta forma :(
