--- Roadmap/32 - BATALLA DEADPOOL Y WOLVERINE/java/armenta-angel/src/main/java/es/armenta/roadmap/Hero.java
@@ -1,19 +1,0 @@
-package es.armenta.roadmap;
-
-public enum Hero {
-
-    DEADPOOL("Deadpool", 10, 100, 25),
-    WOLVERINE("Wolverine", 10, 120, 20);
-
-    final String name;
-    final int minDamage;
-    final int maxDamage;
-    final int defendAttackPercentage;
-
-    Hero(String name, int minDamage, int maxDamage, int defendAttackPercentage) {
-        this.name = name;
-        this.minDamage = minDamage;
-        this.maxDamage = maxDamage;
-        this.defendAttackPercentage = defendAttackPercentage;
-    }
-}
